requirements.txt
appdirs==1.4.4
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
black==19.10b0
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.5
colorlog==6.7.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
djangorestframework==3.13.1
drf-yasg==1.21.3
har2case==0.3.1
httprunner==3.1.6
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.9.5
loguru==0.4.1
MarkupSafe==2.1.1
more-itertools==8.14.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==0.13.1
py==1.11.0
pydantic==1.8.2
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==2.0.2
pytz==2022.2.1
PyYAML==5.4.1
regex==2022.8.17
requests==2.28.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==0.14.4
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
win32-setctime==1.1.0
